+++ app/auth/__init__.py
@@ -1,0 +1,1 @@
+from .views import auth_bp as auth_bp
