+++ app/auth/templates/email/reset_password.html
@@ -1,12 +1,12 @@
 <p>{{ _('Hi %(username)s,', username=user.username) }}</p>
 <p>
     {{ _('Reset your password by clicking the this link:') }}
-    <a href="{{ url_for('user.reset_password', token=token, _external=True) }}">
+    <a href="{{ url_for('auth.reset_password', token=token, _external=True) }}">
         {{ _('click here') }}
     </a>.
 </p>
 <p>{{ _('Alternatively, paste the following link in your browser\'s address bar:') }}</p>
-<p>{{ url_for('user.reset_password', token=token, _external=True) }}</p>
+<p>{{ url_for('auth.reset_password', token=token, _external=True) }}</p>
 <p>{{ _('If you have not requested a password reset, simply ignore this message.') }}</p>
 <p>{{ _('Sincerely,') }}</p>
