+++ app/auth/templates/email/reset_password.txt
@@ -2,7 +2,7 @@ Dear {{ user.username }},
 
 {{ _('Reset your password by clicking the this link:') }}
 
-{{ url_for('user.reset_password', token=token, _external=True) }}
+{{ url_for('auth.reset_password', token=token, _external=True) }}
 
 {{  _('If you have not requested a password reset, simply ignore this message.') }}
