+++ app/public/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import views  # noqa
+from .views import public_bp
