+++ app/user/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 """The user module."""
-from . import views  # noqa
+from .views import user_bp
