+++ babel.cfg
@@ -1,6 +1,7 @@
 [python: app/**.py]
 [jinja2: app/templates/**.html]
-[jinja2: app/templates/email/**.html]
+[jinja2: app/auth/**.html]
+[jinja2: app/auth/email/**.html]
 [jinja2: app/public/templates/**.html]
 [jinja2: app/user/templates/**.html]
 extensions=jinja2.ext.autoescape,jinja2.ext.with_
