+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-addopts = --verbose -rsxX --showlocals --tb=short
+addopts = --verbose -rsxX --showlocals --tb=short --cov=app
 xfail_strict=true
